code-for-30-days/D1_internetSpeed/origin/requirements.txt
speedtest-cli==2.1.2
Flask==1.1.1